Aliases no longer needed
I have rewritten the code so now is possible to add a new column by
$grid->addColumn("p_published"); which will be translated into
p.published.

Only thing you have to do is to specify primary ID (used by grid for
massive actions and row editation) in constructor() in doctrine style so
for example p_id. This is required by the Grid itself.